// src/app/crime-class-information-investigation-core/models/enum.ts
export enum Controls {
    btnSave = 'btnSave',
    btnAdvoAssi = 'btnAdvoAssi',
    tbDSCCRef = 'tbDSCCRef',
    dtpCloseDate = 'dtpCloseDate',
    dtpBillDate = 'dtpBillDate',
}

export enum ModelProperty {
    additionalFixedFee = 'additionalFixedFee',
    additionalWorkFee = 'additionalWorkFee',
    billedDate = 'billedDate',
    branchId = 'branchId',
    cdsDirectTAFChecked = 'cdsDirectTAFChecked',
    claimDisbursements = 'claimDisbursements',
    claimProfitCost = 'claimProfitCost',
    claimTravel = 'claimTravel',
    claimWaiting = 'claimWaiting',
    classDisbursements = 'classDisbursements',
    classFeeEarnTotal = 'classFeeEarnTotal',
    classId = 'classId',
    classProfitCost = 'classProfitCost',
    classTravel = 'classTravel',
    classWaiting = 'classWaiting',
    closedDate = 'closedDate',
    disability = 'disability',
    doNotClaimVATChecked = 'doNotClaimVATChecked',
    dsccRef = 'dsccRef',
    dutySoliChecked = 'dutySoliChecked',
    escapeAmount = 'escapeAmount',
    ethnicCode = 'ethnicCode',
    excludeFromCDS6Checked = 'excludeFromCDS6Checked',
    excludeFromCDS6Enable = 'excludeFromCDS6Enable',
    fileId = 'fileId',
    fixedFee = 'fixedFee',
    gender = 'gender',
    matterType = 'matterType',
    noOfDefendants = 'noOfDefendants',
    nonFixedStageReached = 'nonFixedStageReached',
    openedDate = 'openedDate',
    outcomeCode = 'outcomeCode',
    policeAdvCalls = 'policeAdvCalls',
    policeStId = 'policeStId',
    policeStName = 'policeStName',
    schemeId = 'schemeId',
    schemeName = 'schemeName',
    stageReached = 'stageReached',
    subClassMode = 'subClassMode',
    youthChecked = 'youthChecked',
    invClassCurrentTotalsViewModel = 'invClassCurrentTotalsViewModel'
}

export enum Message {
    closeMSG =
    'This bill is now classed as Escape and Fixed fee will be paid until it is assessed. Please complete the extra forms required.'
}